refactor 등수 이넘 클래스 BiPredicate 활용으로 인한 불필요한 코드 정리
