+++ src/test/java/lotto/domain/place/RankTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.place;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
